Copy descriptor set
After submission a descriptor becomes dirty meaning it cannot be changed. The copy creates a new one that can be changed until the descriptor has been bound in a command buffer.

The old descriptor is scheduled for cleanup and will be removed once all pending command buffers have finished executing